+++ src/components/home/Deposits.jsx
@@ -7,21 +7,21 @@ function preventDefault(event) {
   event.preventDefault();
 }
 
-export default function Deposits() {
-  return (
-    <>
-      <Title>Recent Deposits</Title>
-      <Typography component="p" variant="h4">
-        $3,024.00
-      </Typography>
-      <Typography color="text.secondary" sx={{ flex: 1 }}>
-        on 15 March, 2019
-      </Typography>
-      <div>
-        <Link color="primary" href="/" onClick={preventDefault}>
-          View balance
-        </Link>
-      </div>
-    </>
-  );
-}
+const Deposits = () => (
+  <>
+    <Title>Recent Deposits</Title>
+    <Typography component="p" variant="h4">
+      $3,024.00
+    </Typography>
+    <Typography color="text.secondary" sx={{ flex: 1 }}>
+      on 15 March, 2019
+    </Typography>
+    <div>
+      <Link color="primary" href="/" onClick={preventDefault}>
+        View balance
+      </Link>
+    </div>
+  </>
+);
+
+export default Deposits;
