+++ src/components/home/MainListItems.jsx
@@ -1,0 +1,38 @@
+import React from "react";
+import ListItemButton from "@mui/material/ListItemButton";
+import ListItemIcon from "@mui/material/ListItemIcon";
+import ListItemText from "@mui/material/ListItemText";
+import DashboardIcon from "@mui/icons-material/Dashboard";
+import PeopleIcon from "@mui/icons-material/People";
+import BookmarksIcon from "@mui/icons-material/Bookmarks";
+import ThumbsUpDown from "@mui/icons-material/ThumbsUpDown";
+
+const MainListItems = () => (
+  <>
+    <ListItemButton>
+      <ListItemIcon>
+        <DashboardIcon />
+      </ListItemIcon>
+      <ListItemText primary="Dashboard" />
+    </ListItemButton>
+    <ListItemButton>
+      <ListItemIcon>
+        <BookmarksIcon />
+      </ListItemIcon>
+      <ListItemText primary="Favorite Locations" />
+    </ListItemButton>
+    <ListItemButton>
+      <ListItemIcon>
+        <PeopleIcon />
+      </ListItemIcon>
+      <ListItemText primary="Users" />
+    </ListItemButton>
+    <ListItemButton>
+      <ListItemIcon>
+        <ThumbsUpDown />
+      </ListItemIcon>
+      <ListItemText primary="Reviews" />
+    </ListItemButton>
+  </>
+);
+export default MainListItems;
