+++ src/shared/variables.js
@@ -1,0 +1,2 @@
+const drawerWidth = 240;
+export default drawerWidth;
