+++ src/theme.js
@@ -2,6 +2,9 @@
 import { createTheme } from "@material-ui/core/styles";
 
 const theme = createTheme({
+  palette: {
+    mode: "light",
+  },
   breakpoints: {
     values: {
       xs: 500,
